pyproject.toml
[project]
name = "langgph-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core (>=0.3.63,<0.4.0)",
    "langchain-openai (>=0.3.18,<0.4.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)",
    "langgraph (>=0.4.7,<0.5.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "time-machine (>=2.16.0,<3.0.0)",
    "tavily-python (>=0.7.3,<0.8.0)"
]

[tool.poetry]
packages = [{include = "langgph_project", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
